fix(web): keep clarification controls on one row
